+++ doc-user-mkdocs-project/docs/appendix-release-notes/release-notes.md
@@ -15,8 +15,12 @@ Plugin release notes:
 
 **Feature release to add the `AwsS3Catalog` command.**
 
+* ![bug](bug.png) [1.1.0] Update the [`AwsCloudFront`](../command-ref/AwsCloudFront/AwsCloudFront.md) command:
+    + Fix issue where waiting for invalidation to complete was not occurring.
 * ![new](new.png) [1.1.0] Add the [`AwsS3Catalog`](../command-ref/AwsS3Catalog/AwsS3Catalog.md) command to
   generate landing pages for S3 datasets.
+* ![change](change.png) [1.1.0] Update the [`AwsS3`](../command-ref/AwsS3/AwsS3.md) command:
+    + Allow the file pattern for upload to contain wildcards.
 
 ## Version 1.0.0 ##